requirements.txt
opencv-python
google-cloud-vision>=2.2.0
google-cloud-translate>=3.0.2
requests>=2.3.0
pysrt>=1.0.1
progressbar2>=3.34.3
six>=1.11.0
paddleocr>=2.0.0
paddlepaddle==2.0.0
google-cloud-texttospeech==2.3.0